removed DeletePostButton
