--- client/src/components/DeletePostButton/DeletePostButton.jsx
@@ -1,32 +1,0 @@
-import PropTypes from "prop-types";
-import Button from "../Button.jsx";
-import useFetch from "../../hooks/useFetch";
-import { useContext } from "react";
-import StatusContext from "../../context/status/StatusContext.js";
-
-function DeletePostButton({ postId, onDelete, className }) {
-  const { isLoading } = useContext(StatusContext);
-  const { performFetch } = useFetch(`/post/${postId}`, () => {
-    if (onDelete) onDelete();
-  });
-
-  const handleDelete = () => {
-    if (!postId) return;
-    if (!window.confirm("Delete post?")) return;
-    performFetch({ method: "DELETE" });
-  };
-
-  return (
-    <Button onClick={handleDelete} className={className} disabled={isLoading}>
-      Delete post
-    </Button>
-  );
-}
-
-DeletePostButton.propTypes = {
-  postId: PropTypes.string.isRequired,
-  onDelete: PropTypes.func,
-  className: PropTypes.string,
-};
-
-export default DeletePostButton;
