+++ client/src/components/PostFooter/PostFooter.jsx
@@ -5,9 +5,9 @@ import PropTypes from "prop-types";
 import { CommentIcon, ShareIcon, MoreIcon } from "../icons/index.js";
 import LikeButton from "../LikeButton/LikeButton.jsx";
 import { useNavigate } from "react-router-dom";
-import DeletePostButton from "../DeletePostButton/DeletePostButton.jsx";
 import { useContext } from "react";
 import UserContext from "../../context/user/UserContext.js";
+import useFetch from "../../hooks/useFetch.js";
 
 function PostFooter({ postId, tags, authorId }) {
   const { user } = useContext(UserContext);
@@ -15,6 +15,9 @@ function PostFooter({ postId, tags, authorId }) {
   const [menuOpen, setMenuOpen] = useState(false);
   const menuRef = useRef(null);
   const navigate = useNavigate();
+  const { performFetch } = useFetch(`/post/${postId}`, () => {
+    setMenuOpen(false);
+  });
 
   useEffect(() => {
     function handleClickOutside(e) {
@@ -33,6 +36,11 @@ function PostFooter({ postId, tags, authorId }) {
   const handleEdit = () => {
     navigate(`/post/${postId}/edit`);
     setMenuOpen(false);
+  };
+  const handleDelete = () => {
+    if (!postId) return;
+    if (!window.confirm("Delete post?")) return;
+    performFetch({ method: "DELETE" });
   };
 
   return (
@@ -66,11 +74,9 @@ function PostFooter({ postId, tags, authorId }) {
               <Button className={style.menuBtn} onClick={handleEdit}>
                 Edit post
               </Button>
-              <DeletePostButton
-                postId={postId}
-                className={style.menuBtn}
-                onDelete={() => setMenuOpen(false)}
-              />
+              <Button className={style.menuBtn} onClick={handleDelete}>
+                Delete post
+              </Button>
             </div>
           )}
         </span>
